pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "smart_support_agent"
version = "0.1.0"
authors = [
  { name="Nilotpal Saha", email="contact.techoneel@gmail.com" },
]
description = "A flexible and extensible customer support automation system"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

dependencies = [
    "langchain",
    "langchain-text-splitters",
    "faiss-cpu",
    "sentence-transformers",
    "tiktoken",
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "requests-mock",
    "pypdf",
    "requests-html",
]

[tool.setuptools]
packages = ["core", "cli", "config", "factory"]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:.*numpy.core._multiarray_umath.*:DeprecationWarning",
]

[tool.coverage.run]
source = ["core", "cli", "config", "factory"]
omit = [
    "*/__init__.py",
    "*/interface/*_handler.py:15:27",  # Ignore abstract method bodies
]
